rework the deprecation story
