+++ akka-cluster-sharding/src/main/resources/reference.conf
@@ -98,14 +98,29 @@ akka.cluster.sharding {
   # Default value of 2 leaves last maximum 2*`snapshot-after` events and 3 snapshots (2 old ones + latest snapshot)
   keep-nr-of-batches = 2
 
-  # The ShardAllocationStrategy to use, if not defined programatically.
-  # Possible values are "least-shard-allocation-strategy" and "least-shard-allocation-strategy2".
-  # "least-shard-allocation-strategy2" is recommended from Akka 2.6.10 and will become the default
-  # in future version. For backwards compatibility the default is still "least-shard-allocation-strategy".
-  allocation-strategy = "least-shard-allocation-strategy"
-
-  # Setting for akka.cluster.sharding.ShardCoordinator.LeastShardAllocationStrategy
+  # Setting for akka.cluster.sharding.LeastShardAllocationStrategy
+  # and akka.cluster.sharding.ShardCoordinator.LeastShardAllocationStrategy
+  #
+  # A new rebalance algorithm was included in Akka 2.6.10. It can reach optimal balance in
+  # less rebalance rounds (typically 1 or 2 rounds). The amount of shards to rebalance in each
+  # round can still be limited to make it progress slower. For backwards compatibility
+  # the new algorithm is not enabled by default. Enable the new algorithm by setting
+  # `rebalance-absolute-limit` > 0, for example:
+  # akka.cluster.sharding.least-shard-allocation-strategy.rebalance-absolute-limit=20
+  # The new algorithm is recommended and will become the default in future versions of Akka.
   least-shard-allocation-strategy {
+    # Maximum number of shards that will be rebalanced in one rebalance round.
+    # The lower of this and `rebalance-relative-limit` will be used.
+    rebalance-absolute-limit = 0
+
+    # Maximum number of shards that will be rebalanced in one rebalance round.
+    # Fraction of total number of (known) shards.
+    # The lower of this and `rebalance-absolute-limit` will be used.
+    rebalance-relative-limit = 0.1
+
+    # Deprecated: Use rebalance-absolute-limit and rebalance-relative-limit instead. This property is not used
+    # when rebalance-absolute-limit > 0.
+    #
     # Threshold of how large the difference between most and least number of
     # allocated shards must be to begin the rebalancing.
     # The difference between number of shards in the region with most shards and
@@ -118,20 +133,11 @@ akka.cluster.sharding {
     # on different nodes before rebalance will occur.
     rebalance-threshold = 1
 
+    # Deprecated: Use rebalance-absolute-limit and rebalance-relative-limit instead. This property is not used
+    # when rebalance-absolute-limit > 0.
+    #
     # The number of ongoing rebalancing processes is limited to this number.
     max-simultaneous-rebalance = 3
-  }
-
-  # Setting for akka.cluster.sharding.ShardCoordinator.LeastShardAllocationStrategy2
-  least-shard-allocation-strategy2 {
-    # Maximum number of shards that will be rebalanced in one rebalance round.
-    # The lower of this and `relative-limit` will be used.
-    absolute-limit = 100
-
-    # Maximum number of shards that will be rebalanced in one rebalance round.
-    # Fraction of total number of (known) shards.
-    # The lower of this and `absolute-limit` will be used.
-    relative-limit = 0.1
   }
 
   external-shard-allocation-strategy {
