+++ akka-cluster-sharding/src/main/scala/akka/cluster/sharding/ClusterSharding.scala
@@ -12,6 +12,8 @@ import scala.collection.immutable
 import scala.concurrent.Await
 import scala.util.control.NonFatal
 
+import com.github.ghik.silencer.silent
+
 import akka.actor.Actor
 import akka.actor.ActorRef
 import akka.actor.ActorSystem
@@ -113,11 +115,11 @@ import akka.util.ccompat.JavaConverters._
  *
  * '''Shard Allocation''':
  * The logic deciding which shards to rebalance is defined in a plugable shard allocation
- * strategy. The default implementation [[ShardCoordinator.LeastShardAllocationStrategy]]
+ * strategy. The default implementation [[ShardCoordinator.LeastShardAllocationStrategy]] or
+ * [[akka.cluster.sharding.LeastShardAllocationStrategy]]
  * picks shards for handoff from the `ShardRegion` with highest number of previously allocated shards.
  * They will then be allocated to the `ShardRegion` with lowest number of previously allocated shards,
- * i.e. new members in the cluster. There is a configurable `rebalance-threshold` of how large
- * the difference must be to begin the rebalancing. This strategy can be replaced by an application
+ * i.e. new members in the cluster. This strategy can be replaced by an application
  * specific implementation.
  *
  * '''Recovery''':
@@ -172,7 +174,6 @@ object ClusterSharding extends ExtensionId[ClusterSharding] with ExtensionIdProv
  */
 class ClusterSharding(system: ExtendedActorSystem) extends Extension {
   import ClusterShardingGuardian._
-  import ShardCoordinator.LeastShardAllocationStrategy
   import ShardCoordinator.ShardAllocationStrategy
 
   private val log = Logging(system, this.getClass)
@@ -655,20 +656,20 @@ class ClusterSharding(system: ExtendedActorSystem) extends Extension {
   }
 
   /**
-   * The default `ShardAllocationStrategy` is configured by `akka.cluster.sharding.allocation-strategy`.
+   * The default `ShardAllocationStrategy` is configured by `least-shard-allocation-strategy` properties.
    */
+  @silent("deprecated")
   def defaultShardAllocationStrategy(settings: ClusterShardingSettings): ShardAllocationStrategy = {
-    settings.tuningParameters.allocationStrategy match {
-      case LeastShardAllocationStrategy2.ConfigValue =>
-        val absoluteLimit = settings.tuningParameters.leastShardAllocation2AbsoluteLimit
-        val relativeLimit = settings.tuningParameters.leastShardAllocation2RelativeLimit
-        new LeastShardAllocationStrategy2(absoluteLimit, relativeLimit)
-      case LeastShardAllocationStrategy.ConfigValue =>
-        val threshold = settings.tuningParameters.leastShardAllocationRebalanceThreshold
-        val maxSimultaneousRebalance = settings.tuningParameters.leastShardAllocationMaxSimultaneousRebalance
-        new LeastShardAllocationStrategy(threshold, maxSimultaneousRebalance)
-      case other =>
-        throw new IllegalArgumentException(s"Unknown allocation-strategy: [$other]")
+    if (settings.tuningParameters.leastShardAllocationAbsoluteLimit > 0) {
+      // new algorithm
+      val absoluteLimit = settings.tuningParameters.leastShardAllocationAbsoluteLimit
+      val relativeLimit = settings.tuningParameters.leastShardAllocationRelativeLimit
+      new akka.cluster.sharding.LeastShardAllocationStrategy(absoluteLimit, relativeLimit)
+    } else {
+      // deprecated algorithm
+      val threshold = settings.tuningParameters.leastShardAllocationRebalanceThreshold
+      val maxSimultaneousRebalance = settings.tuningParameters.leastShardAllocationMaxSimultaneousRebalance
+      new ShardCoordinator.LeastShardAllocationStrategy(threshold, maxSimultaneousRebalance)
     }
   }
 }
